+++ CHANGELOG.md
@@ -1,16 +1,35 @@
 Changelog for paco.models
 =========================
 
-7.5.0 (unreleased)
+7.5.1 (unreleased)
 ------------------
 
 ### Added
+
+- Added a `add_stack_hooks` to `paco.models.base.Resource`.
+
+
+7.5.0 (2020-09-17)
+------------------
+
+### Added
+
+- Added `paco.models.registry` as a place to contain configuration that extends or changes Paco.
 
 - Added `IIAMUserResource` as an application-level IAMUser resource.
 
 - Minimal `IPinpointApplication` schema for AWS Pinpoint support.
 
+- AlarmSets and CWLogging are loaded into `project.monitor`. These are used by `paco describe` feature.
+
+- Added `extend_base_schema` hook to the loader to allows Services to extend schemas before the loader loads.
+
+- Container loader can load empty objects (objects with no fields, only a name)
+
 ### Changed
+
+- `paco.modes.services.list_service_plugins` changed to `list_enabled_services`. Returns ony enabed services
+  in a dict format.
 
 - `IIAMUserProgrammaticAccess` changed to `IEnablable` and now defaults to True.